Improving code quality ...

+++ src/main/java/jmetal/experiments/MainC.java
@@ -137,7 +137,7 @@ public class MainC {
 
       if (algorithm.getOutputParameter("evaluations") != null) {
         Integer evals = (Integer) algorithm.getOutputParameter("evaluations");
-        int evaluations = (Integer) evals.intValue();
+        int evaluations = (Integer) evals;
         logger_.info("Speed      : " + evaluations + " evaluations");
       }
     }
