+++ src/main/java/jmetal/metaheuristics/mocell/sMOCell1.java
@@ -77,10 +77,10 @@ public class sMOCell1 extends Algorithm {
     Distance distance = new Distance();
 
     //Read the params
-    populationSize = ((Integer) getInputParameter("populationSize")).intValue();
-    archiveSize = ((Integer) getInputParameter("archiveSize")).intValue();
-    maxEvaluations = ((Integer) getInputParameter("maxEvaluations")).intValue();
-    feedBack = ((Integer) getInputParameter("feedBack")).intValue();
+    populationSize = (Integer) getInputParameter("populationSize");
+    archiveSize = (Integer) getInputParameter("archiveSize");
+    maxEvaluations = (Integer) getInputParameter("maxEvaluations");
+    feedBack = (Integer) getInputParameter("feedBack");
 
 
     //Read the operators
