+++ src/main/java/jmetal/metaheuristics/smsemoa/SMSEMOA_main.java
@@ -151,7 +151,7 @@ public class SMSEMOA_main {
       logger_.info("Spread     : " + indicators.getSpread(population));
       logger_.info("Epsilon    : " + indicators.getEpsilon(population));
 
-      int evaluations = ((Integer) algorithm.getOutputParameter("evaluations")).intValue();
+      int evaluations = (Integer) algorithm.getOutputParameter("evaluations");
       logger_.info("Speed      : " + evaluations + " evaluations");
     }
   }
